requirements.txt
faiss-cpu
llama_cpp_python
wxPython
tqdm
flask
gradio
qdrant-client